Complete PRP-13: Metrics and Statistics Collection
Implemented comprehensive metrics collection system with Prometheus export:

- Created MetricsRegistry with stream, recording, application, and pipeline metrics
- Added MetricsCollector for periodic collection and event handling
- Implemented stream-specific metrics (frames processed, bitrate, FPS, latency)
- Added recording metrics (segments, bytes written, duration, errors)
- Created Prometheus export endpoint at /api/v1/metrics/prometheus
- Added configurable collection intervals and metric toggles
- Integrated metrics collector into API layer with proper lifecycle management
- Added comprehensive test coverage for all metrics functionality
- All tests passing with proper validation gates

Stream managers can now monitor performance, track resource usage, and export metrics to Prometheus for observability.

+++ apps/stream-manager/src/api/dto.rs
@@ -92,7 +92,7 @@ pub enum HealthStatus {
 #[derive(Debug, Clone, Serialize, Deserialize)]
 pub struct SystemMetrics {
     pub cpu_usage_percent: f64,
-    pub memory_usage_mb: u64,
+    pub memory_usage_mb: f64,
     pub disk_usage_gb: f64,
     pub network_throughput_mbps: f64,
 }
